Load serial number from system property when setting device ID
On some devices (e.g. ASUS Zenfone) the serial number of the device
is stored in a file in the /factory partition. Consequently, the only
way to load it is using a user space application during the init
process, and to update the ro.serialno system property.

In most places, TWRP already uses the system property to get
the serial number of the device. However, when generating the
device ID used for the backup folder name it checks the boot
command line directly. On these devices, the serialno is not
included on the command line.

Check the ro.serialno system property additionally to make
sure we get the correct one on these devices.

Change-Id: I4a2eeca883f38fdaeb1209507fd8ebe44b1b04a0